feat: implement local web server for API key retrieval flow
